Restructured simulation and adapted test for linear function.
