+++ ruspy/simulation/simulation.py
@@ -7,13 +7,11 @@ relevant variables.
 import numpy as np
 import pandas as pd
 
-from ruspy.model_code.cost_functions import calc_obs_costs
-from ruspy.model_code.cost_functions import lin_cost
 from ruspy.simulation.simulation_auxiliary import get_unobs_data
 from ruspy.simulation.simulation_auxiliary import simulate_strategy
 
 
-def simulate(init_dict, ev_known, trans_mat, shock=None):
+def simulate(init_dict, ev_known, costs, trans_mat, shock=None):
     """
     The main function to simulate a decision process in the theoretical framework of
     John Rust's 1987 paper. It reads the inputs from the initiation dictionary and
@@ -54,15 +52,11 @@ def simulate(init_dict, ev_known, trans_mat, shock=None):
     num_buses = init_dict["buses"]
     beta = init_dict["beta"]
     num_periods = init_dict["periods"]
-    params = np.array(init_dict["params"])
-    maint_func = lin_cost  # For now just set this to a linear cost function
     if ev_known.shape[0] != trans_mat.shape[0]:
         raise ValueError(
             "The transition matrix and the expected value of the agent "
             "need to have the same size."
         )
-    num_states = ev_known.shape[0]
-    costs = calc_obs_costs(num_states, maint_func, params, 0.001)
     maint_shock_dist_name, repl_shock_dist_name, loc_scale = get_unobs_data(shock)
     states, decisions, utilities = simulate_strategy(
         num_periods,
